--- a/.gitignore
+++ b/.gitignore
@@ -149,3 +149,5 @@ docs/source/_generated/
 # sid
 .sid/
 src/__init__.py
+
+