--- a/.gitignore
+++ b/.gitignore
@@ -149,5 +149,3 @@ docs/source/_generated/
 # sid
 .sid/
 src/__init__.py
-
-