--- a/environment.yml
+++ b/environment.yml
@@ -37,11 +37,6 @@ dependencies:
   - sphinx-copybutton
   - sphinx-autoapi
 
-  # for bokeh png export
-  - selenium
-  - geckodriver
-  - firefox
-
   # Jupyter Lab Formatter
   - black
   - jupyterlab_code_formatter