--- a/src/sid/contacts.py
+++ b/src/sid/contacts.py
@@ -110,10 +110,8 @@ def calculate_infections_by_contacts(
         indexers_list.append(ind)
 
     np.random.seed(next(seed))
-    loop_entries = _get_loop_entries(len(states), len(indexers))
 
-    indices = np.random.choice(len(loop_entries), replace=False, size=len(loop_entries))
-    loop_order = loop_entries[indices]
+    loop_order = _get_shuffled_loop_entries(len(states), len(indexers), next(seed))
 
     (
         infected,
@@ -148,21 +146,36 @@ def calculate_infections_by_contacts(
 
 
 @nb.njit
-def _get_loop_entries(n_states, n_contact_models):
+def _get_shuffled_loop_entries(n_states, n_contact_models, seed, n_model_orders=1000):
     """Create an array of loop entries.
 
-    Examples:
-        >>> _get_loop_entries(2, 2)
-        array([[0, 0],
-               [0, 1],
-               [1, 0],
-               [1, 1]]...
+    We save the loop entries of the following loop in shuffled order:
 
-    """
-    res = np.empty((n_states * n_contact_models, 2), dtype=np.int64)
-    counter = 0
     for i in range(n_states):
         for j in range(n_contact_models):
+            pass
+
+    The loop entries are stored in an array with n_states * n_contact_model rows
+    and two columns. The first column contains the i, the second the j elements.
+
+    Achieving complete randomness would require us to first store all loop entries
+    in an array and then shuffle it. However, this would be very slow. Instead
+    we loop over states in random order and cycle through previously
+
+    """
+    np.random.seed(seed)
+    res = np.empty((n_states * n_contact_models, 2), dtype=np.int64)
+    shuffled_state_indices = np.random.choice(n_states, replace=False, size=n_states)
+
+    # create random permutations of the model orders
+    model_orders = np.zeros((n_model_orders, n_contact_models))
+    for m in range(n_model_orders):
+        model_orders[m] = np.random.choice(n_contact_models, replace=False, size=n_contact_models)
+
+    counter = 0
+
+    for i in shuffled_state_indices:
+        for j in model_orders[i % n_model_orders]:
             res[counter, 0] = i
             res[counter, 1] = j
             counter += 1