--- a/tests/test_simulate.py
+++ b/tests/test_simulate.py
@@ -22,8 +22,7 @@ CONTACT_MODELS = {
 }
 
 
-@pytest.mark.parametrize("debug", [True, False])
-def test_simple_run(params, initial_states, tmp_path, debug):
+def test_simple_run(params, initial_states, tmp_path):
     initial_infections = pd.Series(index=initial_states.index, data=False)
     initial_infections.iloc[0] = True
 
@@ -33,7 +32,6 @@ def test_simple_run(params, initial_states, tmp_path, debug):
         initial_infections,
         CONTACT_MODELS,
         path=tmp_path,
-        debug=debug,
     )
 
     df = simulate(params)