# Description

- baseline: 240 Seconds
- reverse loop order when we calculate infections from recurrent contacts: 200 seconds
- faster shuffling of individual-contact_model combinations: 144 seconds
- faster drawing of random contacts in non-recurrent contact models: 126 seconds
- improvements in update states: 102 seconds
- save reduced dataset and further improvements to update_states: 76 seconds

@tobiasraabe, I suggest you go along this list when making the review.